Move base listing classes to opengever.tabbedview.
